Create landingPage as replacement for HomePage Create ExerciseBrowser as replacement for MainPage Change app routing as needed for the above changes
