+++ frontend/src/pages/exercise_browser/ExerciseBrowserProps.ts
@@ -1,0 +1,6 @@
+import {Exercise} from "../../App.tsx";
+
+export type ExerciseBrowserProps = {
+    exercises: Exercise[];
+    fetchData: () => void;
+}
