+++ frontend/src/pages/landingPage/LandingPage.css
@@ -1,0 +1,84 @@
+.centered {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.flex-col {
+    flex-direction: column;
+}
+
+.flex-row {
+    flex-direction: row;
+}
+
+.flex-1 {
+    flex: 1;
+}
+
+.flex-2 {
+    flex: 2;
+}
+
+.w-45 {
+    width: 45%;
+}
+
+.w-80 {
+    width: 80%;
+}
+
+.w-100 {
+    width: 100%;
+}
+
+.min-w-300 {
+    min-width: 300px;
+}
+
+.min-h-150 {
+    min-height: 150px;
+}
+
+.h-80 {
+    height: 80vh;
+}
+
+.pl-10 {
+    padding-left: 10%;
+}
+
+.pr-10 {
+    padding-right: 10%;
+}
+
+.m-10 {
+    margin: 10px;
+}
+
+.mb-5vh {
+    margin-bottom: 5vh;
+}
+
+.direction-col-row-row{
+    flex-direction: column;
+
+    @media (min-width: 650px) {
+        flex-direction: row;
+    }
+}
+
+.direction-col {
+    flex-direction: column;
+}
+
+.scaled-img {
+    object-fit: cover;
+    width: 50vw;
+    max-height: 256px;
+    max-width: 256px;
+
+    @media (min-width: 650px) {
+        width: 35vw;
+    }
+}
